mouse-server/requirements.txt
pyautogui-0.9.54
websockets-12.0